Send SMS while running background
백그라운드 실행 중 서버로부터 데이터를 받아 SMS 전송
